refactor(diffshub): flatten UI components
Move shared UI wrappers out of components/ui and rename their files to match their exported React components.

Update DiffsHub imports and the local shadcn alias so new UI component paths resolve from the top-level components directory.
